Breaking out the route tests so can use setup as common test class. Adding a means to send different response on empty partial observation
